+++ graal/com.oracle.graal.hotspot.amd64/src/com/oracle/graal/hotspot/amd64/AMD64HotSpotLIRGenerator.java
@@ -78,6 +78,7 @@ public class AMD64HotSpotLIRGenerator extends AMD64LIRGenerator implements HotSp
         assert config.basicLockSize == 8;
         this.config = config;
         this.stub = stub;
+        memoryPeephole = new AMD64HotSpotMemoryPeephole(this);
     }
 
     @Override