+++ graal/com.oracle.graal.nodes/src/com/oracle/graal/nodes/calc/CompareNode.java
@@ -26,6 +26,7 @@ import com.oracle.graal.api.meta.*;
 import com.oracle.graal.graph.*;
 import com.oracle.graal.graph.spi.*;
 import com.oracle.graal.nodes.*;
+import com.oracle.graal.nodes.extended.*;
 import com.oracle.graal.nodes.spi.*;
 
 /* TODO (thomaswue/gdub) For high-level optimization purpose the compare node should be a boolean *value* (it is currently only a helper node)
@@ -34,7 +35,7 @@ import com.oracle.graal.nodes.spi.*;
  * Compare should probably be made a value (so that it can be canonicalized for example) and in later stages some Compare usage should be transformed
  * into variants that do not materialize the value (CompareIf, CompareGuard...)
  */
-public abstract class CompareNode extends LogicNode implements Canonicalizable, LIRLowerable {
+public abstract class CompareNode extends LogicNode implements Canonicalizable, LIRLowerable, MemoryArithmeticLIRLowerable {
 
     @Input private ValueNode x;
     @Input private ValueNode y;
@@ -191,5 +192,9 @@ public abstract class CompareNode extends LogicNode implements Canonicalizable, 
         }
 
         return graph.unique(comparison);
+    }
+
+    public boolean generate(MemoryArithmeticLIRLowerer gen, Access access) {
+        return false;
     }
 }