+++ graal/com.oracle.graal.nodes/src/com/oracle/graal/nodes/calc/ReinterpretNode.java
@@ -26,6 +26,7 @@ import com.oracle.graal.api.meta.*;
 import com.oracle.graal.graph.*;
 import com.oracle.graal.graph.spi.*;
 import com.oracle.graal.nodes.*;
+import com.oracle.graal.nodes.extended.*;
 import com.oracle.graal.nodes.spi.*;
 import com.oracle.graal.nodes.type.*;
 
@@ -34,7 +35,7 @@ import com.oracle.graal.nodes.type.*;
  * of a primitive value to some other incompatible stamp. The new stamp must have the same width as
  * the old stamp.
  */
-public class ReinterpretNode extends FloatingNode implements Canonicalizable, ArithmeticLIRLowerable {
+public class ReinterpretNode extends FloatingNode implements Canonicalizable, ArithmeticLIRLowerable, MemoryArithmeticLIRLowerable {
 
     @Input private ValueNode value;
 
@@ -104,6 +105,15 @@ public class ReinterpretNode extends FloatingNode implements Canonicalizable, Ar
     public void generate(ArithmeticLIRGenerator gen) {
         PlatformKind kind = gen.getPlatformKind(stamp());
         gen.setResult(this, gen.emitReinterpret(kind, gen.operand(value())));
+    }
+
+    @Override
+    public boolean generate(MemoryArithmeticLIRLowerer gen, Access access) {
+        Value result = gen.emitReinterpretMemory(stamp(), access);
+        if (result != null) {
+            gen.setResult(this, result);
+        }
+        return result != null;
     }
 
     public static ValueNode reinterpret(Kind toKind, ValueNode value) {